move workfiles related logic to workfiles model
